+++ Fabric/src/main/java/fuzs/puzzleslib/client/core/FabricClientRegistration.java
@@ -1,7 +1,12 @@
 package fuzs.puzzleslib.client.core;
 
+import fuzs.puzzleslib.client.init.builder.ModSpriteParticleRegistration;
+import net.fabricmc.fabric.api.client.particle.v1.ParticleFactoryRegistry;
 import net.fabricmc.fabric.api.client.rendering.v1.TooltipComponentCallback;
 import net.minecraft.client.gui.screens.inventory.tooltip.ClientTooltipComponent;
+import net.minecraft.client.particle.ParticleProvider;
+import net.minecraft.core.particles.ParticleOptions;
+import net.minecraft.core.particles.ParticleType;
 import net.minecraft.world.inventory.tooltip.TooltipComponent;
 
 import java.util.function.Function;
@@ -18,5 +23,15 @@ public class FabricClientRegistration implements ClientRegistration {
             if (data.getClass() == type) return factory.apply((T) data);
             return null;
         });
+    }
+
+    @Override
+    public <T extends ParticleOptions> void registerParticleProvider(ParticleType<T> type, ParticleProvider<T> provider) {
+        ParticleFactoryRegistry.getInstance().register(type, provider);
+    }
+
+    @Override
+    public <T extends ParticleOptions> void registerParticleProvider(ParticleType<T> type, ModSpriteParticleRegistration<T> factory) {
+        ParticleFactoryRegistry.getInstance().register(type, factory::create);
     }
 }